+++ labelbox/data/annotation_types/geometry/mask.py
@@ -2,8 +2,9 @@ from typing import Callable, Optional, Tuple, Union, Dict, List
 
 import numpy as np
 from pydantic.class_validators import validator
-from shapely.geometry import MultiPolygon, Polygon
 import cv2
+
+from shapely.geometry import MultiPolygon, Polygon
 
 from ..data import MaskData
 from .geometry import Geometry